Move ApacheCXF feature to root of cxf package
