+++ binding/cxf/src/main/java/io/tracee/binding/cxf/TraceeCxfFeature.java
@@ -1,4 +1,4 @@
-package io.tracee.binding.cxf.client;
+package io.tracee.binding.cxf;
 
 import io.tracee.Tracee;
 import io.tracee.TraceeBackend;
