+++ binding/cxf/src/test/java/io/tracee/binding/cxf/CxfClientToJaxwsServerIT.java
@@ -1,7 +1,6 @@
 package io.tracee.binding.cxf;
 
 import io.tracee.*;
-import io.tracee.binding.cxf.client.TraceeCxfFeature;
 import io.tracee.binding.cxf.testSoapService.HelloWorldTestService;
 import io.tracee.binding.jaxws.client.TraceeClientHandler;
 import org.apache.cxf.feature.LoggingFeature;
